Multimodel_AI/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
python-dotenv==1.0.1
google-generativeai==0.8.3
slowapi==0.1.9
pydantic[email]==2.10.4
aiosmtplib==3.0.2
tenacity==9.0.0
redis==5.2.1
loguru==0.7.3
